+++ tests/test_cli.py
@@ -1,11 +1,47 @@
-from nose.tools import *
 from phone2address import cli
+import requests
+import openpyxl
+
+with open('correct_response.html', 'r', encoding='gbk') as f:
+    corrent_data = f.readlines()
+    f.close()
+
+with open('wrong_response.html', 'r') as f:
+    wrong_data = f.readlines()
+    f.close()
+
+response = requests.Response
 
 
-def test_get_address():
-    assert_equal(cli.get_address('13012345678'), ['重庆', ''])
-    assert_equal(cli.get_address('110'), False)
+def mock_request_get(url):
+    base = 'http://www.ip138.com:8080/search.asp?action=mobile&mobile='
+    if url == base + '13012345678':
+        response.text = str(corrent_data)
+        return response
+    else:
+        response.text = str(wrong_data)
+        return response
 
 
-def test_process_xlsx():
-    pass
+def read_from_file(source):
+    from itertools import chain
+    tmp = list(chain.from_iterable([i for i in source]))
+    data = []
+    for i in tmp:
+        if i.value is not None:
+            data.append(str(i.value))
+    return data
+
+
+def test_get_address(monkeypatch):
+    monkeypatch.setattr('requests.get', mock_request_get)
+    assert cli.get_address('13012345678') == ['重庆', '']
+    assert cli.get_address('110') == False
+
+
+def test_process_xlsx(monkeypatch):
+    monkeypatch.setattr('requests.get', mock_request_get)
+    cli.process('test_xlsx.xlsx')
+    source = openpyxl.load_workbook('test_processed.xlsx').active
+    target = openpyxl.load_workbook('processed.xlsx').active
+    assert read_from_file(source) == read_from_file(target)